working on dialogs
Intermediary commit before testing


Fixed refactor bugs


inline form works again


Dialog seems to work


dialog works
