--- lib/django-form/django-form.component.html
@@ -1,6 +1,0 @@
-
-
-<ng-template #dynamicComponentContainer>
-</ng-template>
-
-<md-progress-bar mode="indeterminate" *ngIf="loading"></md-progress-bar>
