+++ lib/django-form/django-form.module.ts
@@ -1,19 +1,19 @@
 import {NgModule} from '@angular/core';
-import {DjangoFormComponent} from './django-form.component';
 import {DynamicFormsCoreModule} from '@ng-dynamic-forms/core/src/core.module';
 import {DynamicFormsMaterialUIModule} from '@ng-dynamic-forms/ui-material/src/dynamic-material-form-ui.module';
 import {NG_VALIDATORS, ReactiveFormsModule} from '@angular/forms';
 import {CommonModule} from '@angular/common';
 import {InPageDjangoFormComponent} from './impl/inpage-django-form.component';
-import {MdButtonModule, MdProgressBarModule, MdSnackBarModule} from '@angular/material';
+import {MdButtonModule, MdDialogModule, MdProgressBarModule, MdSnackBarModule} from '@angular/material';
 import {MdButtonGroupModule} from '../material/button-group/button.group.module';
-import {external_validator, InternalDjangoFormContentComponent} from './impl/internal-django-form-content.component';
+import {DjangoFormContentComponent, external_validator} from './impl/django-form-content.component';
+import {DialogDjangoFormComponent} from './impl/dialog-django-form.component';
 
 @NgModule({
     declarations: [
-        DjangoFormComponent,
         InPageDjangoFormComponent,
-        InternalDjangoFormContentComponent
+        DjangoFormContentComponent,
+        DialogDjangoFormComponent,
     ],
     imports: [
         CommonModule,
@@ -24,15 +24,18 @@ import {external_validator, InternalDjangoFormContentComponent} from './impl/int
         MdButtonModule,
         MdButtonGroupModule,
         MdSnackBarModule,
+        MdDialogModule
     ],
     providers: [
         {provide: NG_VALIDATORS, useValue: external_validator, multi: true}
     ],
     exports: [
-        DjangoFormComponent
+        InPageDjangoFormComponent,
+        DialogDjangoFormComponent,
     ],
     entryComponents: [
-        InPageDjangoFormComponent
+        InPageDjangoFormComponent,
+        DialogDjangoFormComponent,
     ]
 })
 export class DjangoFormModule {
