+++ lib/django-form/impl/dialog-django-form.component.html
@@ -1,0 +1,18 @@
+<h1 md-dialog-title *ngIf="form_title">{{ form_title }}</h1>
+<div md-dialog-content>
+    <django-form-content [layout]="layout" [initial_data]="initial_data" [errors]="errors"
+                                  (submit)="submitted()" #form>
+    </django-form-content>
+</div>
+
+<div md-dialog-actions>
+    <button
+            md-raised-button
+            *ngFor="let button of actions"
+            [name]="button.id"
+            [color]="button.color"
+            (click)="submitted(button.id, button.cancel)"
+            type="button"
+            [disabled]="!form.valid && !button.cancel"
+    >{{ button.label }}</button>
+</div>
