+++ lib/django-form/impl/dialog-django-form.component.scss
@@ -1,0 +1,3 @@
+h1 {
+    margin-bottom: 20px;
+}
