+++ lib/django-form/impl/inpage-django-form.component.ts
@@ -1,17 +1,16 @@
-import {Component, EventEmitter, Input, OnInit, Output, ViewChild} from '@angular/core';
-import {InternalDjangoFormBaseComponent} from './internal-django-form-base.component';
+import {Component} from '@angular/core';
+import {Http} from '@angular/http';
+import {MdSnackBar} from '@angular/material';
+import {DjangoFormBaseComponent} from './django-form-base.component';
 
-/**
- * Form component targeted on django rest framework
- */
 @Component({
     selector: 'inpage-django-form',
     templateUrl: './inpage-django-form.component.html',
     styleUrls: ['./inpage-django-form.component.scss'],
 })
-export class InPageDjangoFormComponent extends InternalDjangoFormBaseComponent {
+export class InPageDjangoFormComponent extends DjangoFormBaseComponent {
 
-    constructor() {
-        super();
+    constructor(http: Http, snackBar: MdSnackBar) {
+        super(http, snackBar);
     }
 }
