--- lib/django-form/impl/internal-django-form-content.component.html
@@ -1,6 +1,0 @@
-<form [formGroup]="form_group" (ngSubmit)="onSubmit(null)" (keydown.enter)="onSubmit()">
-
-  <dynamic-material-form [group]="form_group"
-                         [model]="form_model"></dynamic-material-form>
-
-</form>
