+++ lib/django-form/index.ts
@@ -1,1 +1,4 @@
 export * from './django-form.module';
+export * from './django-form-dialog.service';
+export * from './impl/inpage-django-form.component';
+export * from './impl/dialog-django-form.component';
